simplified snippet retrieval (includes major changes!)
The plugin uses a new API functionality now to only get an extract instead
of the whole text of an article. That also means that a few previously
available things had to be removed, but a few new things have been added
as well and the changes improve the code.

* images and tables are not supported by the extract functionality; that's why the two config options `includeImages` and `includeTables` have been removed as well
* the html parser (simplehtmldom) has been removed because it's not necessary anymore
* permalinks have been removed, because the article's revision is not available
* two new config options have been added:
  - `snippetLength` let's you define the number of sentences in a snippet (empty/default: show intro section)
  - `useHtml` let's you choose if the snippet should include HTML or be plain text